API Reference

Thunks module

Thinkers.Thunk — Type

Thunk(callable, args...; kwargs...)

Hold a callable and its arguments for lazy evaluation. Use reify! to evaluate.

Examples

julia> using Thinkers

julia> a = Thunk(x -> 3x, 4);

julia> reify!(a);

julia> getresult(a)
Some(12)

julia> b = Thunk(+, 4, 5);

julia> reify!(b);

julia> getresult(b)
Some(9)

julia> c = Thunk(sleep, 1);

julia> getresult(c)  # `c` has not been evaluated

julia> reify!(c);  # `c` has been evaluated

julia> getresult(c)
Some(nothing)

julia> f(args...; kwargs...) = collect(kwargs);

julia> d = Thunk(f, 1, 2, 3; x=1.0, y=4, z="5");

julia> reify!(d);

julia> getresult(d)
Some(Pair{Symbol, Any}[:x => 1.0, :y => 4, :z => "5"])

julia> e = Thunk(sin, "1");  # Catch errors

julia> reify!(e);

julia> haserred(e)
true

Thinkers.ErrorInfo — Type

Capture errors and stack traces from a running Thunk.

Thinkers.reify! — Function

reify!(thunk::Thunk)

Reify a Thunk into a value.

Compute the value of the expression. Walk through the Thunk's arguments and keywords, recursively evaluating each one, and then evaluating the Thunk's function with the evaluated arguments.

See also Thunk.

Thinkers.isevaluated — Function

isevaluated(thunk::Thunk)

Determine whether thunk has been executed before.

Thinkers.haserred — Function

haserred(thunk::Thunk)

Check if thunk produced an error when executed.